moving brick functions to a seperate file
